silly little component working beautifully
