+++ src/pages/AccountPage/AccountPage.tsx
@@ -1,14 +1,16 @@
-import { useEffect } from 'react';
+import { useEffect, useState } from 'react';
 
 import { useSearchParams } from 'react-router-dom';
 
+import { SkillComboBox } from '../../components/SkillComboBox/SkillComboBox';
 import { useIsCoLinksSite } from '../../features/colinks/useIsCoLinksSite';
 import { ShowOrConnectGitHub } from '../../features/github/ShowOrConnectGitHub';
 import { ShowOrConnectLinkedIn } from '../../features/linkedin/ShowOrConnectLinkedIn';
 import { ShowOrConnectTwitter } from '../../features/twitter/ShowOrConnectTwitter';
 import { useToast } from '../../hooks';
+import { X } from '../../icons/__generated';
 import { EditEmailForm } from 'pages/ProfilePage/EmailSettings/EditEmailForm';
-import { ContentHeader, Flex, Panel, Text } from 'ui';
+import { ContentHeader, Flex, IconButton, Panel, Text } from 'ui';
 import { SingleColumnLayout } from 'ui/layouts';
 
 import { EditProfileInfo } from './EditProfileInfo';
@@ -40,6 +42,8 @@ export default function AccountPage() {
         </Flex>
       </ContentHeader>
       <Flex column css={{ maxWidth: '$readable', gap: '$lg' }}>
+        Boom
+        <PickOneSkill />
         {isCoLinksPage && (
           <>
             <Panel css={{ gap: '$md' }}>
@@ -85,3 +89,33 @@ export default function AccountPage() {
     </SingleColumnLayout>
   );
 }
+
+const PickOneSkill = () => {
+  const [skill, setSkill] = useState<string | undefined>();
+  return (
+    <>
+      {skill ? (
+        <Flex css={{ gap: '$md' }}>
+          <Text tag size="medium" color="complete" css={{ pr: 0 }}>
+            {skill}
+            <IconButton
+              onClick={() => setSkill(undefined)}
+              css={{ pr: '$sm', width: 'auto' }}
+            >
+              <X size={'xs'} />
+            </IconButton>
+          </Text>
+        </Flex>
+      ) : (
+        <SkillComboBox
+          hideInput={false}
+          excludeSkills={[]}
+          addSkill={async (skill: string) => {
+            setSkill(skill);
+          }}
+          placeholder={'Choose a GIVE Reason'}
+        />
+      )}
+    </>
+  );
+};
