+++ src/pages/AccountPage/SkillAndTopicPicker.tsx
@@ -1,29 +1,19 @@
 import assert from 'assert';
-import { useRef, useState } from 'react';
 
-import * as Popover from '@radix-ui/react-popover';
-import { Command, useCommandState } from 'cmdk';
 import { useMutation, useQuery, useQueryClient } from 'react-query';
 
-import { ComboBox } from '../../components/ComboBox';
 import { LoadingIndicator } from '../../components/LoadingIndicator';
-import useProfileId from '../../hooks/useProfileId';
-import { User, X } from '../../icons/__generated';
-import { order_by, skills_constraint } from '../../lib/gql/__generated__/zeus';
-import { client } from '../../lib/gql/client';
 import {
-  Flex,
-  IconButton,
-  Panel,
-  PopoverContent,
-  Text,
-  TextField,
-} from '../../ui';
+  QUERY_KEY_ALL_SKILLS,
+  SkillComboBox,
+} from '../../components/SkillComboBox/SkillComboBox';
+import useProfileId from '../../hooks/useProfileId';
+import { X } from '../../icons/__generated';
+import { skills_constraint } from '../../lib/gql/__generated__/zeus';
+import { client } from '../../lib/gql/client';
+import { Flex, IconButton, Panel, Text } from '../../ui';
 
 const QUERY_KEY_PROFILE_SKILLS = 'profile_skills';
-const QUERY_KEY_ALL_SKILLS = 'skills';
-
-const MAX_POTENTIAL_SKILLS = 1000;
 
 const insertSkill = async (skill: string) => {
   await client.mutate(
@@ -47,36 +37,6 @@ const insertSkill = async (skill: string) => {
       operationName: 'add_skill',
     }
   );
-};
-
-// TODO: maybe this should do server-side filtering but prob not needed for awhile
-// that's why there is the commented out where clause.
-const fetchSkills = async () => {
-  const { skills } = await client.query(
-    {
-      skills: [
-        {
-          // where: query
-          //   ? {
-          //       name: {
-          //         _ilike: '%' + query + '%',
-          //       },
-          //     }
-          //   : undefined,
-          order_by: [{ count: order_by.desc }, { name: order_by.asc }],
-          limit: MAX_POTENTIAL_SKILLS,
-        },
-        {
-          name: true,
-          count: true,
-        },
-      ],
-    },
-    {
-      operationName: 'fetchPotentialSkills',
-    }
-  );
-  return skills;
 };
 
 const fetchMySkills = async (profileId: number) => {
@@ -164,7 +124,6 @@ export const SkillAndTopicPicker = () => {
     {
       onSuccess: async () => {
         await queryClient.invalidateQueries([QUERY_KEY_PROFILE_SKILLS]);
-        setPopoverOpen(false);
       },
     }
   );
@@ -231,168 +190,12 @@ export const SkillAndTopicPicker = () => {
               <SkillComboBox
                 hideInput={maxedOut}
                 excludeSkills={profileSkills}
+                addSkill={async (skill: string) => addSkillToProfile(skill)}
               />
             </Flex>
           </Flex>
         </>
       )}
     </Panel>
-  );
-};
-
-type SkillComboBoxProps = {
-  hideInput: boolean;
-  excludeSkills: string[];
-};
-const SkillComboBox = ({ hideInput, excludeSkills }: SkillComboBoxProps) => {
-  const { data: skills, isLoading } = useQuery(
-    [QUERY_KEY_ALL_SKILLS],
-    fetchSkills
-  );
-
-  const inputRef = useRef<HTMLInputElement>(null);
-  const [popoverOpen, setPopoverOpen] = useState<boolean>(false);
-
-  if (!skills || isLoading) {
-    return <LoadingIndicator />;
-  }
-  return (
-    <Popover.Root open={popoverOpen} onOpenChange={setPopoverOpen}>
-      {!hideInput && (
-        <Flex
-          column
-          as={Popover.Trigger}
-          css={{
-            alignItems: 'flex-start',
-            gap: '$sm',
-            borderRadius: '$3',
-          }}
-        >
-          {/* This TextField is just a popover trigger */}
-          <TextField
-            placeholder="Search or Add Interest"
-            css={{ width: '302px' }}
-            value=""
-          />
-        </Flex>
-      )}
-      <PopoverContent
-        avoidCollisions={false}
-        align={'start'}
-        css={{
-          background: 'transparent',
-          mt: 'calc(-$1xl + 0.5px)',
-          p: 0,
-        }}
-      >
-        <ComboBox
-          filter={(value, search) => {
-            if (value == search.toLowerCase()) {
-              return 1;
-            } else if (value.includes(search.toLowerCase())) return 0.9;
-            return 0;
-          }}
-        >
-          <Command.Input
-            ref={inputRef}
-            placeholder={'Search or Add Interest'}
-            maxLength={30}
-          />
-
-          <Command.List>
-            {isLoading ? (
-              <Command.Loading>LoadingMate</Command.Loading>
-            ) : (
-              <>
-                <AddItem
-                  addSkill={addSkill}
-                  alreadyAddedSkills={Array.from(excludeSkills)}
-                  allSkills={skills}
-                />
-
-                <Command.Group>
-                  {skills
-                    .filter(
-                      sk =>
-                        !excludeSkills.some(
-                          ps => ps.toLowerCase() === sk.name.toLowerCase()
-                        )
-                    )
-                    .map(skill => (
-                      <Command.Item
-                        key={skill.name}
-                        value={skill.name}
-                        onSelect={addSkill}
-                        defaultChecked={false}
-                        disabled={excludeSkills.some(
-                          ps => ps.toLowerCase() === skill.name.toLowerCase()
-                        )}
-                      >
-                        <Flex
-                          css={{
-                            justifyContent: 'space-between',
-                            width: '100%',
-                          }}
-                        >
-                          <Text semibold>{skill.name}</Text>
-                          <Text tag color={'secondary'} size={'xs'}>
-                            <User /> {skill.count}
-                          </Text>
-                        </Flex>
-                      </Command.Item>
-                    ))}
-                </Command.Group>
-              </>
-            )}
-          </Command.List>
-        </ComboBox>
-      </PopoverContent>
-    </Popover.Root>
-  );
-};
-
-type Skill = Awaited<ReturnType<typeof fetchSkills>>[number];
-
-const AddItem = ({
-  addSkill,
-  alreadyAddedSkills,
-  allSkills,
-}: {
-  addSkill(skill: string): void;
-  alreadyAddedSkills: string[];
-  allSkills: Skill[];
-}) => {
-  const search = useCommandState(state => state.search);
-  if (
-    search.trim() === '' ||
-    allSkills.some(s => s.name.toLowerCase() === search.toLowerCase())
-  ) {
-    return null;
-  }
-
-  if (alreadyAddedSkills.some(s => s.toLowerCase() === search.toLowerCase())) {
-    return (
-      <Command.Item color={'cta'} key={search} value={search} disabled={true}>
-        <Text semibold>Already added {search}</Text>
-      </Command.Item>
-    );
-  }
-
-  return (
-    <Command.Item
-      color={'cta'}
-      key={search}
-      value={search}
-      onSelect={() => addSkill(search)}
-    >
-      <Flex
-        css={{ justifyContent: 'space-between', width: '100%', gap: '$lg' }}
-      >
-        <Text>Be the First to Add</Text>
-        <Text tag color={'complete'} semibold size={'medium'}>
-          {search}
-        </Text>
-      </Flex>
-    </Command.Item>
   );
 };
